WIP #8487 #7877
